Move SectionDetail to DT package
